requirements.txt
Flask
Flask-Cors
pandas
openpyxl
numpy
google-generativeai
langchain-core
gunicorn
os